requirements.txt
Django==2.2.16
django-debug-toolbar==2.2
django-js-asset==2.0.0
django-mptt==0.13.4
mixer==7.1.2
Pillow==8.3.1
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
requests==2.26.0
six==1.16.0
sorl-thumbnail==12.7.0
Faker==12.0.1
